fix: Move middleware.ts to src/ so Next.js actually loads it
The root middleware.ts was never picked up by Next.js because this
project uses a src/ directory. Moved to src/middleware.ts and updated
the import path from './src/i18n/config' to './i18n/config'.
Updated test import accordingly.
